backend/requirements.txt
Flask
Flask-CORS
gunicorn
GitPython
requests